pyproject.toml
[tool.poetry]
name = "product-comparison"
version = "0.1.0"
description = ""
authors = ["Tarik El Lel <tareklel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
